refactor: move embedding file into fetchers directory
