+++ lib/fetchers/embedding.ts
