+++ lib/fetchers/index.ts
@@ -2,4 +2,5 @@ export * from "./search";
 export * from "./blog";
 export * from "./courses";
 export * from "./classes";
-export * from "./parents"
+export * from "./parents";
+export * from "./embedding";
